requirements.txt
absl-py==2.1.0
accelerate==0.34.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
binpacking==1.5.2
bitsandbytes==0.43.3
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
dacite==1.8.1
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==5.3.1b1
editdistance==0.8.1
einops==0.8.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evalscope==0.5.4
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
fastjsonschema==2.20.0
feedparser==6.0.11
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
func_timeout==4.3.5
future==1.0.0
fuzzywuzzy==0.18.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.0
gradio_client==1.3.0
griffe==1.3.1
grpcio==1.66.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hf_transfer==0.1.8
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.8
imageio==2.35.1
immutabledict==4.2.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
interegular==0.3.3
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lagent==0.2.3
lark==1.2.2
-e git+https://githubfast.com/QwenLM/Qwen2.5-Math.git@e128aa06871c7a0eeedc2ab21b69459bcb24c4fb#egg=latex2sympy2&subdirectory=evaluation/latex2sympy
lazy_loader==0.4
Levenshtein==0.26.0
llvmlite==0.43.0
lm-format-enforcer==0.10.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.4.4
mistune==3.0.2
mmengine==0.10.5
mmengine-lite==0.10.5
modelscope==1.18.1
mpmath==1.3.0
ms-opencompass==0.1.0
-e git+https://github.com/modelscope/swift.git@c3ac7c1459194bb5a26912fd90e600caa6082148#egg=ms_swift
ms-vlmeval==0.0.7
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
narwhals==1.8.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.43.0
OpenCC==1.1.9
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.7
oss2==2.19.0
outlines==0.0.46
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
Pebble==5.0.7
peft==0.12.0
pexpect==4.9.0
phx-class-registry==4.0.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
ply==3.11
portalocker==2.10.1
prettytable==3.11.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
Pympler==1.1
pyparsing==3.1.4
pypinyin==0.53.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.26.0
python-multipart==0.0.10
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
RapidFuzz==3.10.0
ray==2.37.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.20.0
ruff==0.6.7
sacrebleu==2.4.3
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.1.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
shtab==1.7.1
simple-ddl-parser==1.6.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.6
streamlit==1.38.0
sty==1.0.6
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.7.0
timeout-decorator==0.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.3.0
torchaudio==2.4.1+cu121
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.42.3
transformers-stream-generator==0.0.5
triton==2.3.0
trl==0.10.1
typer==0.12.5
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
unsloth @ git+https://github.com/unslothai/unsloth.git@d91d40a7b6b556f2d1fdd3e1e430f7a76a799627
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
validators==0.34.0
vllm==0.5.1
vllm-flash-attn==2.5.9
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
word2number==1.1
xformers==0.0.26.post1
XlsxWriter==3.2.0
xtuner==0.1.23
xxhash==3.5.0
yapf==0.40.2
yarl==1.9.11
zipp==3.20.1
